requirements.txt
# NexusFile AI - Dependencies
flask==3.1.0
flask-cors==5.0.1
google-genai==1.14.0
python-pptx==1.0.2
python-docx==1.1.2
openpyxl==3.1.5
reportlab==4.4.0
PyMuPDF==1.25.4
Pillow==11.1.0
chardet==5.2.0
SQLAlchemy==2.0.36
python-dotenv==1.0.1
Werkzeug==3.1.3
